back/requirements.txt

djangorestframework>=3.15.2
django-cors-headers==4.3.1
PyJWT>=2.0
python-dotenv>=1.0